# docs/swagger.yaml
definitions:
...
  model_base.BaseArticleLibrary:
    properties:
      IsOpen:
        description: 是否公开库
        type: string
      libraryId:
        description: 库id
        type: integer
      libraryName:
        description: 库名
        type: string
      libraryUrl:
        description: 库的路径
        type: string
      status:
        description: 状态
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
  model_sys.SysDictData:
    properties:
      cssClass:
        description: 样式属性（其他样式扩展）
        type: string
      dictCode:
        description: 字典编码
        type: integer
      dictLabel:
        description: 字典标签
        type: string
      dictSort:
        description: 字典排序
        type: integer
      dictType:
        description: 字典类型
        type: string
      dictValue:
        description: 字典键值
        type: string
      isDefault:
        description: 是否默认（Y是 N否）
        type: string
      listClass:
        description: 表格回显样式
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    type: object
...
  result.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /base/baseArticleLibrary/{libraryId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: libraryId
        in: path
        name: libraryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取库详情
      tags:
      - Base - 文章库管理
  /base/baseArticleLibrary/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticleLibrary
        in: body
        name: baseArticleLibrary
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticleLibrary'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加库
      tags:
      - Base - 文章库管理
  /base/baseArticleLibrary/deletes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: libraryIds
        in: query
        name: libraryIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除库（实体删除）
      tags:
      - Base - 文章库管理
  /base/baseArticleLibrary/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticleLibrary
        in: body
        name: baseArticleLibrary
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticleLibrary'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改库
      tags:
      - Base - 文章库管理
  /base/baseArticleLibrary/myPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 是否公开库
        in: query
        name: IsOpen
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 库id
        in: query
        name: libraryId
        type: integer
      - description: 库名
        in: query
        name: libraryName
        type: string
      - description: 库的路径
        in: query
        name: libraryUrl
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取当前用户的库列表
      tags:
      - Base - 文章库管理
  /base/baseArticleLibrary/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 是否公开库
        in: query
        name: IsOpen
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 库id
        in: query
        name: libraryId
        type: integer
      - description: 库名
        in: query
        name: libraryName
        type: string
      - description: 库的路径
        in: query
        name: libraryUrl
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取库列表
      tags:
      - Base - 文章库管理
  /base/baseArticleLibrary/removes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: libraryIds
        in: query
        name: libraryIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除库（逻辑删除）
      tags:
      - Base - 文章库管理
...
  /sys/sysDict/data/{dictCode}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: dictCode
        in: path
        name: dictCode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取字典数据详情
      tags:
      - System - 字典管理
  /sys/sysDict/data/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: sysDictData
        in: body
        name: sysDictData
        required: true
        schema:
          $ref: '#/definitions/model_sys.SysDictData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加菜单
      tags:
      - System - 字典管理
  /sys/sysDict/data/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: sysDictData
        in: body
        name: sysDictData
        required: true
        schema:
          $ref: '#/definitions/model_sys.SysDictData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改菜单
      tags:
      - System - 字典管理
  /sys/sysDict/data/list/{dictType}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: dictType
        in: path
        name: dictType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取指定字典数据列表
      tags:
      - System - 字典管理
  /sys/sysDict/data/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 样式属性（其他样式扩展）
        in: query
        name: cssClass
        type: string
      - description: 字典编码
        in: query
        name: dictId
        type: integer
      - description: 字典标签
        in: query
        name: dictLabel
        type: string
      - description: 字典排序
        in: query
        name: dictSort
        type: integer
      - description: 字典类型
        in: query
        name: dictType
        type: string
      - description: 字典键值
        in: query
        name: dictValue
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 是否默认（Y是 N否）
        in: query
        name: isDefault
        type: string
      - description: 表格回显样式
        in: query
        name: listClass
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态（0正常 1停用）
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取字典数据列表
      tags:
      - System - 字典管理
...
  /sys/sysDict/data/removes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: dictIds
        in: query
        name: dictIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除菜单[逻辑删除]
      tags:
      - System - 管理
...
swagger: "2.0"